code done now need to complte docs

+++ UncleYellow/YellowHotelDK/YellowHotelDK_Front/src/app/components/admin/admin/dialog-delete-record-any/dialog-delete-record-any.component.ts
@@ -1,5 +1,8 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { EmployeesService } from 'src/app/services/employees.service';
+import { LetterService } from 'src/app/services/letter.service';
+import { RoomBookingService } from 'src/app/services/room-booking.service';
 import { UsersService } from 'src/app/services/users.service';
 import Swal from 'sweetalert2';
 
@@ -13,7 +16,11 @@ export class DialogDeleteRecordAnyComponent implements OnInit {
   constructor(
     public dialogRef: MatDialogRef<DialogDeleteRecordAnyComponent>,
     @Inject(MAT_DIALOG_DATA) public data: any,
-    public usersService: UsersService
+    public usersService: UsersService,
+    public roomBookingService: RoomBookingService,
+    public employeesService: EmployeesService,
+    public letterService: LetterService,
+
   ) {
 
   }
@@ -22,12 +29,38 @@ export class DialogDeleteRecordAnyComponent implements OnInit {
 
   }
   save() {
-    this.usersService.deleteUser(this.data.item.id)
-      .subscribe((res: any) => {
-        // handle response
-        this.processResponse()
-        this.dialogRef.close()
-      });
+    if (this.data.deleteRoom) {
+      this.roomBookingService.deleteBooking(this.data.item.id)
+        .subscribe((rs: any) => {
+          // handle response
+          this.processResponse()
+          this.dialogRef.close()
+        });
+    }
+    if (this.data.deleteUsers) {
+      this.usersService.deleteUser(this.data.item.id)
+        .subscribe((res: any) => {
+          // handle response
+          this.processResponse()
+          this.dialogRef.close()
+        });
+    }
+    if (this.data.deleteEmployees) {
+      this.employeesService.deleteEmployees(this.data.item.id)
+        .subscribe((res: any) => {
+          // handle response
+          this.processResponse()
+          this.dialogRef.close()
+        });
+    }
+    if (this.data.deleteLetters) {
+      this.letterService.deleteLetters(this.data.item.id)
+        .subscribe((res: any) => {
+          // handle response
+          this.processResponse()
+          this.dialogRef.close()
+        });
+    }
   }
   cancel() {
     this.dialogRef.close()
